separating the UI code
